mam2: rename mam files to message
