+++ mam/v2/mam/channel.h
@@ -88,38 +88,6 @@ void mam_channel_destroy(mam_channel_t *const channel);
 
 retcode_t mam_channels_destroy(mam_channel_t_set_t *const channels);
 
-/**
- * Gets the size of a wrapped channel
- *
- * @return the wrapped channel size
- */
-size_t mam_channel_wrap_size();
-
-/**
- * Wraps a channel into a trits buffer
- *
- * @param spongos A spongos interface
- * @param buffer The buffer
- * @param version The MAM version
- * @param channel_id The channel id
- */
-void mam_channel_wrap(mam_spongos_t *const spongos, trits_t *const buffer,
-                      tryte_t const version, trits_t const channel_id);
-
-/**
- * Unwraps a channel from a trits buffer
- *
- * @param spongos A spongos interface
- * @param buffer The buffer
- * @param version The MAM version
- * @param channel_id The channel id
- *
- * @return a status code
- */
-retcode_t mam_channel_unwrap(mam_spongos_t *const spongos,
-                             trits_t *const buffer, tryte_t *const version,
-                             trits_t channel_id);
-
 size_t mam_channel_serialized_size(mam_channel_t const *const channel);
 
 retcode_t mam_channel_serialize(mam_channel_t const *const channel,
