+++ mam/v2/mam/tests/BUILD
@@ -1,8 +1,8 @@
 cc_test(
-    name = "test_mam",
-    srcs = ["test_mam.c"],
+    name = "test_message",
+    srcs = ["test_message.c"],
     deps = [
-        "//mam/v2/mam",
+        "//mam/v2/mam:message",
         "//mam/v2/test_utils",
         "@unity",
     ],
@@ -22,9 +22,9 @@ cc_test(
     name = "test_endpoint",
     srcs = ["test_endpoint.c"],
     deps = [
-        "//mam/v2/mam",
         "//mam/v2/mam:endpoint",
         "//mam/v2/mam:mam_endpoint_t_set",
+        "//mam/v2/mam:message",
         "//mam/v2/test_utils",
         "@unity",
     ],
@@ -34,9 +34,9 @@ cc_test(
     name = "test_channel",
     srcs = ["test_channel.c"],
     deps = [
-        "//mam/v2/mam",
         "//mam/v2/mam:channel",
         "//mam/v2/mam:mam_channel_t_set",
+        "//mam/v2/mam:message",
         "//mam/v2/test_utils",
         "@unity",
     ],
